requirements.txt
asgiref==3.2.7
autopep8==1.5.2
Django==3.0.7
django-taggit==1.2.0
factory-boy==2.12.0
Faker==4.1.0
flake8==3.8.1
Markdown==3.2.2
mccabe==0.6.1
Pillow
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
django-ninja-extra
django-ninja-jwt
ninja-schema
django-markdownx
Pygments
django-cors-headers